Moved dao and factory interfaces to shared so they could be used by SQLplugin. Tests can now load up the SQLDao factory!
