+++ shared/src/main/java/com/example/shared/interfaces/IDaoFactory.java
@@ -1,4 +1,4 @@
-package database;
+package com.example.shared.interfaces;
 
 public interface IDaoFactory {
     public IGameDao createGameDao();
