+++ shared/src/main/java/com/example/shared/interfaces/IUserDao.java
@@ -1,4 +1,4 @@
-package database;
+package com.example.shared.interfaces;
 
 import com.example.shared.model.User;
